requirements.txt
#
# This file is autogenerated by pip-compile with python 3.11
# To update, run:
#
#    pip-compile
#
appnope==0.1.3
    # via ipython
asgiref==3.5.2
    # via
    #   django
    #   djoser
asttokens==2.1.0
    # via stack-data
attrs==22.1.0
    # via
    #   flake8-bugbear
    #   pytest
autopep8==2.0.0
    # via django-silk
backcall==0.2.0
    # via ipython
black==22.10.0
    # via -r requirements.in
blinker==1.5
    # via nplusone
build==0.9.0
    # via pip-tools
certifi==2022.9.24
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   black
    #   pip-tools
commonmark==0.9.1
    # via rich
coreapi==2.3.3
    # via
    #   djoser
    #   drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
coverage[toml]==6.5.0
    # via pytest-cov
cryptography==38.0.4
    # via social-auth-core
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
dj-database-url==1.0.0
    # via -r requirements.in
django==4.1.3
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-anymail
    #   django-cors-headers
    #   django-extensions
    #   django-silk
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-anymail==8.6
    # via -r requirements.in
django-cors-headers==3.13.0
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
django-silk==5.0.2
    # via -r requirements.in
django-templated-mail==1.1.1
    # via djoser
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==4.7.2
    # via
    #   -r requirements.in
    #   djoser
djoser==2.1.0
    # via -r requirements.in
drf-yasg==1.21.4
    # via -r requirements.in
executing==1.2.0
    # via stack-data
factory-boy==3.2.1
    # via -r requirements.in
faker==15.3.3
    # via factory-boy
flake8==5.0.4
    # via
    #   -r requirements.in
    #   flake8-broken-line
    #   flake8-bugbear
    #   flake8-builtins
    #   flake8-comprehensions
    #   flake8-print
    #   flake8-spellcheck
    #   flake8-tuple
    #   pep8-naming
flake8-broken-line==0.6.0
    # via -r requirements.in
flake8-bugbear==22.10.27
    # via -r requirements.in
flake8-builtins==2.0.1
    # via -r requirements.in
flake8-comprehensions==3.10.1
    # via -r requirements.in
flake8-plugin-utils==1.3.2
    # via
    #   flake8-pytest-style
    #   flake8-return
flake8-print==5.0.0
    # via -r requirements.in
flake8-pytest-style==1.6.0
    # via -r requirements.in
flake8-return==1.2.0
    # via -r requirements.in
flake8-spellcheck==0.28.0
    # via -r requirements.in
flake8-tuple==0.4.1
    # via -r requirements.in
freezegun==1.2.2
    # via pytest-freezegun
gprof2dot==2022.7.29
    # via django-silk
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-yasg
iniconfig==1.1.1
    # via pytest
ipython==8.6.0
    # via -r requirements.in
isort==5.10.1
    # via -r requirements.in
itypes==1.2.0
    # via coreapi
jedi==0.18.2
    # via ipython
jinja2==3.1.2
    # via
    #   coreschema
    #   django-silk
markupsafe==2.1.1
    # via jinja2
matplotlib-inline==0.1.6
    # via ipython
mccabe==0.7.0
    # via flake8
mypy==0.991
    # via -r requirements.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
newrelic==8.4.0
    # via -r requirements.in
nplusone==1.0.0
    # via -r requirements.in
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==21.3
    # via
    #   build
    #   drf-yasg
    #   pytest
parso==0.8.3
    # via jedi
pathspec==0.10.2
    # via black
pep517==0.13.0
    # via build
pep8-naming==0.13.2
    # via -r requirements.in
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.10.0
    # via -r requirements.in
platformdirs==2.5.4
    # via black
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.33
    # via ipython
psycopg2==2.9.5
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.9.1
    # via
    #   autopep8
    #   flake8
    #   flake8-print
pycparser==2.21
    # via cffi
pyflakes==2.5.0
    # via flake8
pygments==2.13.0
    # via
    #   ipython
    #   rich
pyjwt==2.6.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
pyparsing==3.0.9
    # via packaging
pytest==7.2.0
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
    #   pytest-freezegun
    #   pytest-instafail
    #   pytest-randomly
    #   pytest-timeout
pytest-cov==4.0.0
    # via -r requirements.in
pytest-django==4.5.2
    # via -r requirements.in
pytest-freezegun==0.4.2
    # via -r requirements.in
pytest-instafail==0.4.2
    # via -r requirements.in
pytest-randomly==3.12.0
    # via -r requirements.in
pytest-timeout==2.1.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   django-silk
    #   faker
    #   freezegun
python3-openid==3.2.0
    # via social-auth-core
pytz==2022.6
    # via
    #   djangorestframework
    #   drf-yasg
pyupgrade==3.2.2
    # via -r requirements.in
pywatchman==1.4.1
    # via -r requirements.in
requests==2.28.1
    # via
    #   coreapi
    #   django-anymail
    #   django-silk
    #   requests-oauthlib
    #   rollbar
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
rich==12.6.0
    # via -r requirements.in
rollbar==0.16.3
    # via -r requirements.in
ruamel-yaml==0.17.21
    # via drf-yasg
sentry-sdk==1.11.1
    # via -r requirements.in
six==1.16.0
    # via
    #   asttokens
    #   flake8-tuple
    #   nplusone
    #   python-dateutil
    #   rollbar
    #   social-auth-app-django
social-auth-app-django==4.0.0
    # via djoser
social-auth-core==4.3.0
    # via social-auth-app-django
sqlparse==0.4.3
    # via
    #   django
    #   django-silk
stack-data==0.6.2
    # via ipython
tokenize-rt==5.0.0
    # via pyupgrade
tomli==2.0.1
    # via autopep8
traitlets==5.5.0
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.4.0
    # via mypy
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.13
    # via
    #   requests
    #   sentry-sdk
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.38.4
    # via pip-tools
whitenoise==6.2.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
